+++ pyqpp/include/pyqpp/qasm/qasm_bind.hpp
@@ -36,6 +36,9 @@ inline void init_qasm_qasm(py::module_& m) {
     auto py_qasm = m.def_submodule("qasm");
     py_qasm.def("read_from_file", &qpp::qasm::read_from_file,
                 "Get QCircuit representation of OpenQASM circuit");
+
+    py_qasm.def("read_from_string", &qpp::qasm::read_from_string,
+                "Get QCircuit representation of OpenQASM circuit");
 }
 
 #endif /* PYQPP_QASM_QASM_BIND_HPP_ */